Change BlockGroup coordinates to float

+++ src/main/java/tetris/domain/BlockGroup.java
@@ -3,10 +3,10 @@ package tetris.domain;
 public class BlockGroup {
 
     private BlockGroupItem[] blocks;
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
-    public BlockGroup(BlockGroupItem[] blocks, int x, int y) {
+    public BlockGroup(BlockGroupItem[] blocks, float x, float y) {
         this.blocks = blocks;
         this.x = x;
         this.y = y;
@@ -28,19 +28,19 @@ public class BlockGroup {
         return blocks;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
